Create Business_of_ISP
You live in a city where there are N buildings, and E roads connect those buildings. Each road has a distance, and they are two-way roads. You want to start a business in your city where you will provide ISP service. For that reason, you need to connect all buildings with your central wire cable through the roads. As the price of optical fiber is high, you want to reduce the cost as much as possible. If you know the distance of the roads, can you calculate the minimum cost to connect all buildings with your cable? The cost of optical fiber is equal to the distance of any road.

Sample Input:
6 10
0 1 5
0 2 6
0 3 8
0 4 6
0 5 10
3 5 2
5 1 4
1 2 3
2 4 8
4 3 6

Sample Output:
20